Movies, DataObject and config all work together
The config file just sets up the constants for the application. So far, only the info for the DB connections exists.

The DataObject file would be the parent file for all the classes. I think that most of our classes would be used to read data and enter data. This is just an idea but the database $pdo object is written just once and created every time we instantiate an object from a class.

The Movies class defines the genres and reads a genre from a movies table...not sure it will work as I have not tested.

Once we have discussed how we are going to structure our data, we could probably create a new repo and pull the required info in.